Lazy loading is now complete
All the row items are now hoverable, and once they are hovered, they
start loading some dummy data (in this case "their favorite games").
Once it has finished loading, the "loading spinner" will disappear, and
instead the data will be displayed. The column will also be filled with
the retrieved data.

TODO: Improve the text, and perhaps clean the code a bit more, and try
to simplify it as much as possible

StarcounterSamples/KitchenSink#78