Code improvement, as suggested by @miyconst
Instead of scheduling 100 session tasks, it now only schedules one.

And instead of depending on, and comparing to the selected persons name,
the scheduling is now based off of the "isHovered" boolean. That way,
even if there are multiple rows with the same name it will work fine

StarcounterSamples/KitchenSink#78