Changed the text a little
the Lazy loading page was changed a bit, and now starts loading the data
on hover, and displays it when it finishes the loading.

Before, it used to "cancel" the loading as the different rows got
hovered. If the user would leave the hovered person, and the data hadnt
loaded yet, it cancelled.

So I changed the text now, since the previous text was no longer
applicable